Исправлены все замечания, скорректированы тесты
