+++ src/tasks/Subtask.java
@@ -36,23 +36,6 @@ public class Subtask extends Task {
     }
 
     @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-
-        Subtask subtask = (Subtask) o;
-        return epicId.equals(subtask.epicId);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = super.hashCode();
-        result = 31 * result + epicId.hashCode();
-        return result;
-    }
-
-    @Override
     public String toString() {
         return "Subtask{" +
                 "id=" + getId() +
